Updated date.js module with datebuilder export.
